Usar path relativo y reestructura de carpetas

+++ Enviroment/boxplot.py
@@ -2,19 +2,19 @@
 import os
 import time
 from matplotlib import pyplot as plt
+from pathlib import Path 
 
 def boxplot(tipos = None,name= ''):
-    graficas = "C://Users//Víctor//Desktop//TFM//Enviroment//Graficas//"
-    path = "C://Users//Víctor//Desktop//TFM//Enviroment//"
+    current_path = str(Path.cwd()).replace("\\","//")
+
+    graficas = current_path + "//Graficas//"
+
+    path = current_path + "//"
     if tipos is None: 
         tipos = ["Accuracy","Coste","Well_Formed","Validity","Testability","Quality"]
     else:
         tipos = [tipos]
     auc = "//auc.txt"
-    # _, axs = plt.subplots(2, 3,figsize = (15,15))
-
-    # i=0
-    # j=0
 
     for tipo in tipos:
         dir = path + tipo + auc
@@ -27,13 +27,6 @@ def boxplot(tipos = None,name= ''):
         tipoa = tipo
         if tipoa == "Coste": tipoa += " (Líneas)"
         axact.set(ylabel = "Area under the curve (AUC)", title= "AUC " + tipoa)
-        # plt.show()
-        # i += 1
-        # if i == 3:
-        #     j += 1
-        #     i = 0
-        # for l in lines:
-        # plt.savefig(graficas+(time.strftime("%x,%X")).replace(r'/',"-").replace(r':',"-")+".png")
         file.close()
         plt.savefig(graficas+tipo+name+"Box"+".png")
 
@@ -42,14 +35,21 @@ def borra(tipos = None):
         tipos = ["Accuracy","Coste","Well_Formed","Validity","Testability","Quality"]
     else:
         tipos = [tipos]    
-    path = "C://Users//Víctor//Desktop//TFM//Enviroment//"
+
+    current_path = str(Path.cwd()).replace("\\","//")
+
+    path = current_path + "//"
     auc = "//auc.txt"
 
     for tipo in tipos:
         dir = path + tipo + auc
-        os.remove(dir)
+
+        if os.path.isfile(dir):
+            os.remove(dir)
+
+        
 
 if __name__ == '__main__':
-    # borra()
-    boxplot("Coste")
+    borra()
+    # boxplot("Coste")
     pass
